Change test/tsize printf st_size to 64 bit
Fix test/tsize printf to handle 64 bit st_size.
This was causing test/t12 to fail on AIX 5.2.

+++ test/tsize.c
@@ -47,7 +47,7 @@ main(int argc, char *argv[])
         perror(argv[1]);
         exit(1);
     }
-    printf("%ld\n", sb.st_size);
+    printf("%lld\n", (unsigned long long)sb.st_size);
     exit(0);
 }
